Fixes a lot. Did too much in this one commit. Need to work on that.
Design decision for #9 is to KISS and leave out error processing since we
can't detect everything that might be wrong. Move requirements/assumptions
to the comment of method. That is the contract the caller must follow in
order to get meaningful results.
Design decision in #9 makes #8 obsolete. Error checking is moved to
comment of methods to make it KISS. Closes #8.
Cell checks for validity not at initialization like it used to, but when
Cell#valid? is called.
Puzzle#to_s now converts a blank cell's text to a period instead of the
cell itself. This is because that functionality is specific to printing
the puzzle.
Changed more comments to end with a period for consistency. #19
Added some descriptive comments. #7